# backend/api/rest/openapi.yaml
openapi: 3.0.3
info:
  title: OneBook AI Gateway API
  version: 0.1.0
  description: |
    Gateway REST API contract. All public requests go through the gateway.
    If the client sends `X-Request-Id`, the gateway propagates it end-to-end.
    If missing, the gateway generates one and returns it in both response header
    `X-Request-Id` and error payload field `requestId`.
servers:
  - url: http://localhost:8080
    description: Local dev (gateway)
...
paths:
...
  /api/chats:
    post:
      tags: [chat]
      summary: Ask a question about a book
      description: |
        The service may include the most recent conversation history when
        generating the answer (history length is configurable). Obvious follow-up
        questions may be answered from recent conversation history without a new
        retrieval pass, while obvious out-of-scope realtime questions are
        rejected instead of querying the book index.
      security:
        - sessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Book not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/conversations:
    get:
      tags: [chat]
      summary: List conversation history
      description: Returns the current user's latest conversation list.
      security:
        - sessionCookieAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListConversationsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: onebook_access
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Stable machine-readable error code (domain-specific, e.g. AUTH_INVALID_CREDENTIALS / BOOK_NOT_FOUND / CHAT_BOOK_NOT_READY).
        requestId:
          type: string
          description: Request correlation id, same as response header X-Request-Id.
        details:
          type: array
          description: Optional field-level validation details.
          items:
            $ref: "#/components/schemas/ErrorDetail"
      required: [error, code]
      example:
        error: book not ready
        code: CHAT_BOOK_NOT_READY
        requestId: req-demo-20260213-001
        details:
          - reason: book_status_conflict
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
      required: [reason]
...
    Source:
      type: object
      properties:
        label:
          type: string
        location:
          type: string
        snippet:
          type: string
        chunkId:
          type: string
        sourceRef:
          type: string
        score:
          type: number
          format: double
        language:
          type: string
      required: [label, location, snippet]
    Answer:
      type: object
      properties:
        conversation:
          $ref: "#/components/schemas/ConversationSummary"
        question:
          type: string
        answer:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Source"
        abstained:
          type: boolean
        retrievalDebug:
          $ref: "#/components/schemas/RetrievalDebug"
        createdAt:
          type: string
          format: date-time
      required: [conversation, question, answer, citations, abstained, createdAt]
    ChatRequest:
      type: object
      properties:
        conversationId:
          type: string
        bookId:
          type: string
        question:
          type: string
        debug:
          type: boolean
      required: [bookId, question]
    RetrievalHit:
      type: object
      properties:
        chunkId:
          type: string
        sourceRef:
          type: string
        score:
          type: number
          format: double
        stage:
          type: string
        snippet:
          type: string
      required: [chunkId, score, stage]
    RetrievalDebug:
      type: object
      properties:
        language:
          type: string
        queries:
          type: array
          items:
            type: string
        dense:
          type: array
          items:
            $ref: "#/components/schemas/RetrievalHit"
        lexical:
          type: array
          items:
            $ref: "#/components/schemas/RetrievalHit"
        fused:
          type: array
          items:
            $ref: "#/components/schemas/RetrievalHit"
        reranked:
          type: array
          items:
            $ref: "#/components/schemas/RetrievalHit"
      required: [language, queries, dense, lexical, fused, reranked]
    ConversationSummary:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        bookId:
          type: string
        title:
          type: string
        lastMessageAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, title, createdAt, updatedAt]
...
    ListConversationsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ConversationSummary"
        count:
          type: integer
      required: [items, count]
...