# backend/services/chat/internal/server/server.go
package server

import (
	"encoding/json"
	"errors"
	"io"
	"net/http"
	"net/url"
	"strconv"
	"strings"

	"onebookai/internal/usertoken"
	"onebookai/internal/util"
	"onebookai/pkg/domain"
	"onebookai/services/chat/internal/app"
	"onebookai/services/chat/internal/authclient"
	"onebookai/services/chat/internal/bookclient"
)

// Config wires required dependencies for the HTTP server.
type Config struct {
	App           *app.App
	Auth          *authclient.Client
	Books         *bookclient.Client
	TokenVerifier *usertoken.Verifier
}

// Server exposes HTTP endpoints for the chat service.
type Server struct {
	app           *app.App
	auth          *authclient.Client
	books         *bookclient.Client
	tokenVerifier *usertoken.Verifier
	mux           *http.ServeMux
}

// New constructs the server with routes configured.
func New(cfg Config) *Server {
	s := &Server{
		app:           cfg.App,
		auth:          cfg.Auth,
		books:         cfg.Books,
		tokenVerifier: cfg.TokenVerifier,
		mux:           http.NewServeMux(),
	}
	s.routes()
	return s
}

// Router returns the configured handler.
func (s *Server) Router() http.Handler {
	return util.WithRequestID(util.WithRequestLog("chat", util.WithSecurityHeaders(util.WithCORS(s.mux))))
}

func (s *Server) routes() {
	s.mux.HandleFunc("/healthz", s.handleHealth)
	s.mux.Handle("/chats", s.withUser(s.handleChats))
	s.mux.Handle("/conversations", s.withUser(s.handleConversations))
	s.mux.Handle("/conversations/", s.withUser(s.handleConversationByID))
}

func (s *Server) handleHealth(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

type userHandler func(http.ResponseWriter, *http.Request, string, domain.User)

func (s *Server) withUser(next userHandler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if s.auth == nil {
			writeError(w, http.StatusInternalServerError, "auth client not configured")
			return
		}
		token, ok := bearerToken(r)
		if !ok {
			writeError(w, http.StatusUnauthorized, "unauthorized")
			return
		}
		if s.tokenVerifier != nil {
			if _, err := s.tokenVerifier.VerifySubject(token); err != nil {
				writeError(w, http.StatusUnauthorized, "unauthorized")
				return
			}
		}
		user, err := s.auth.Me(token)
		if err != nil {
			writeError(w, http.StatusUnauthorized, "unauthorized")
			return
		}
		next(w, r, token, user)
	})
}

func (s *Server) handleChats(w http.ResponseWriter, r *http.Request, token string, user domain.User) {
	if r.Method != http.MethodPost {
		methodNotAllowed(w)
		return
	}
	var req chatRequest
	if err := json.NewDecoder(io.LimitReader(r.Body, 1<<20)).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "invalid JSON body")
		return
	}
	if req.Question == "" {
		writeError(w, http.StatusBadRequest, "question is required")
		return
	}
	idempotencyKey := util.IdempotencyKeyFromRequest(r)

	// For existing conversations, resolve bookId from the stored conversation
	// so the frontend doesn't need to track which book a conversation belongs to.
	resolvedBookID := strings.TrimSpace(req.BookID)
	if strings.TrimSpace(req.ConversationID) != "" {
		convBookID, err := s.app.GetConversationBookID(user, req.ConversationID)
		if err != nil {
			if errors.Is(err, app.ErrConversationNotFound) {
				writeError(w, http.StatusNotFound, err.Error())
			} else if errors.Is(err, app.ErrConversationForbidden) {
				writeError(w, http.StatusForbidden, err.Error())
			} else {
				writeError(w, http.StatusInternalServerError, err.Error())
			}
			return
		}
		if convBookID != "" {
			resolvedBookID = convBookID
		}
	}

	if resolvedBookID == "" {
		writeError(w, http.StatusBadRequest, "bookId is required")
		return
	}
	if s.books == nil {
		writeError(w, http.StatusInternalServerError, "book client not configured")
		return
	}
	book, err := s.books.GetBook(token, resolvedBookID)
	if err != nil {
		writeBookError(w, err)
		return
	}
	ans, replayed, err := s.app.AskQuestion(user, book, req.Question, req.ConversationID, idempotencyKey, req.Debug && user.Role == domain.RoleAdmin)
	if err != nil {
		status := http.StatusBadRequest
		if errors.Is(err, app.ErrBookNotReady) {
			status = http.StatusConflict
		} else if errors.Is(err, app.ErrConversationNotFound) {
			status = http.StatusNotFound
		} else if errors.Is(err, app.ErrConversationForbidden) {
			status = http.StatusForbidden
		} else if strings.Contains(err.Error(), "forbidden") {
			status = http.StatusForbidden
		}
		writeError(w, status, err.Error())
		return
	}
	if replayed {
		w.Header().Set("Idempotency-Replayed", "true")
	}
	writeJSON(w, http.StatusOK, ans)
}

func (s *Server) handleConversations(w http.ResponseWriter, r *http.Request, _ string, user domain.User) {
	if r.Method != http.MethodGet {
		methodNotAllowed(w)
		return
	}
	limit := readLimitParam(r.URL.Query(), 30, 100)
	items, err := s.app.ListConversations(user, limit)
	if err != nil {
		writeError(w, http.StatusBadRequest, err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]any{
		"items": items,
		"count": len(items),
	})
}

func (s *Server) handleConversationByID(w http.ResponseWriter, r *http.Request, _ string, user domain.User) {
	path := strings.TrimPrefix(r.URL.Path, "/conversations/")
	path = strings.Trim(path, "/")
	if path == "" {
		writeError(w, http.StatusNotFound, "not found")
		return
	}
	parts := strings.Split(path, "/")
	if len(parts) != 2 || parts[1] != "messages" {
		writeError(w, http.StatusNotFound, "not found")
		return
	}
	if r.Method != http.MethodGet {
		methodNotAllowed(w)
		return
	}
	conversationID := parts[0]
	limit := readLimitParam(r.URL.Query(), 200, 500)
	items, err := s.app.ListConversationMessages(user, conversationID, limit)
	if err != nil {
		status := http.StatusBadRequest
		if errors.Is(err, app.ErrConversationNotFound) {
			status = http.StatusNotFound
		} else if errors.Is(err, app.ErrConversationForbidden) {
			status = http.StatusForbidden
		}
		writeError(w, status, err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]any{
		"items": items,
		"count": len(items),
	})
}

func readLimitParam(values url.Values, fallback int, max int) int {
	raw := strings.TrimSpace(values.Get("limit"))
	if raw == "" {
		return fallback
	}
	parsed, err := strconv.Atoi(raw)
	if err != nil || parsed <= 0 {
		return fallback
	}
	if parsed > max {
		return max
	}
	return parsed
}

func methodNotAllowed(w http.ResponseWriter) {
	writeError(w, http.StatusMethodNotAllowed, "method not allowed")
}

type chatRequest struct {
	ConversationID string `json:"conversationId,omitempty"`
	BookID         string `json:"bookId"`
	Question       string `json:"question"`
	Debug          bool   `json:"debug,omitempty"`
}

func writeJSON(w http.ResponseWriter, status int, payload any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(payload)
}

type errorResponse struct {
	Error     string `json:"error"`
	Code      string `json:"code"`
	RequestID string `json:"requestId,omitempty"`
}

func writeError(w http.ResponseWriter, status int, msg string) {
	writeErrorWithCode(w, status, msg, "")
}

func writeErrorWithCode(w http.ResponseWriter, status int, msg, code string) {
	code = strings.TrimSpace(code)
	if code == "" {
		code = errorCodeForChat(status, msg)
	}
	if code == "" {
		code = "REQUEST_ERROR"
	}
	writeJSON(w, status, errorResponse{
		Error:     msg,
		Code:      code,
		RequestID: strings.TrimSpace(w.Header().Get("X-Request-Id")),
	})
}

func writeBookError(w http.ResponseWriter, err error) {
	if apiErr, ok := err.(*bookclient.APIError); ok {
		writeErrorWithCode(w, apiErr.Status, apiErr.Message, apiErr.Code)
		return
	}
	writeErrorWithCode(w, http.StatusBadGateway, "book service unavailable", "BOOK_SERVICE_UNAVAILABLE")
}

func errorCodeForChat(status int, msg string) string {
	message := strings.ToLower(strings.TrimSpace(msg))
	switch {
	case message == "auth client not configured", message == "book client not configured":
		return "SYSTEM_INTERNAL_ERROR"
	case message == "unauthorized":
		return "AUTH_INVALID_TOKEN"
	case message == "question is required":
		return "CHAT_QUESTION_REQUIRED"
	case message == "bookid is required":
		return "CHAT_BOOK_ID_REQUIRED"
	case message == "conversation id required":
		return "CHAT_CONVERSATION_ID_REQUIRED"
	case message == "conversation not found":
		return "CHAT_CONVERSATION_NOT_FOUND"
	case message == "conversation forbidden":
		return "CHAT_CONVERSATION_FORBIDDEN"
	case message == "book not ready":
		return "CHAT_BOOK_NOT_READY"
	case message == "forbidden":
		return "CHAT_FORBIDDEN"
	case message == "invalid json body":
		return "CHAT_INVALID_REQUEST"
	case message == "method not allowed":
		return "SYSTEM_METHOD_NOT_ALLOWED"
	case message == "book service unavailable":
		return "BOOK_SERVICE_UNAVAILABLE"
	}
	switch status {
	case http.StatusBadRequest:
		return "CHAT_INVALID_REQUEST"
	case http.StatusUnauthorized:
		return "AUTH_INVALID_TOKEN"
	case http.StatusForbidden:
		return "CHAT_FORBIDDEN"
	case http.StatusNotFound:
		return "CHAT_CONVERSATION_NOT_FOUND"
	case http.StatusConflict:
		return "CHAT_BOOK_NOT_READY"
	case http.StatusMethodNotAllowed:
		return "SYSTEM_METHOD_NOT_ALLOWED"
	case http.StatusBadGateway, http.StatusServiceUnavailable:
		return "SYSTEM_UPSTREAM_UNAVAILABLE"
	default:
		if status >= http.StatusInternalServerError {
			return "SYSTEM_INTERNAL_ERROR"
		}
		return "REQUEST_ERROR"
	}
}

func bearerToken(r *http.Request) (string, bool) {
	authHeader := r.Header.Get("Authorization")
	if !strings.HasPrefix(authHeader, "Bearer ") {
		return "", false
	}
	token := strings.TrimSpace(strings.TrimPrefix(authHeader, "Bearer "))
	if token == "" {
		return "", false
	}
	return token, true
}